Extract domain checking code to domain checker
We end up using this code in multiple places.
Better to live in the domain checking object
